Add /dev/ptp0 passthrough
Some virtual cloud services provide a very good system time on their
host, which can be accessed by the guest VMs through the `/dev/ptp0`
device. This interface is found on Hyper-V and KVM (the latter
requires the `ptp_kvm` module to be loaded).

In my experience the time is very stable and often quite accurate,
often helping chrony to converge faster to a good time. As a result,
this commit adds an optional functionality to pass the host's
`/dev/ptp0` device to the container.